Extract SignupRepository for user signup logic
- Created `SignupRepository` inside `org.ole.planet.myplanet.lite.signup` package to handle raw Request construction, URL building, and OkHttp executions for checking username availability and submitting signup payloads.
- Refactored `SignupActivity` to utilize `SignupRepository`, moving the network dependencies and logic out of the activity to maintain separation of concerns.
- Migrated tests by renaming `SignupActivityExecuteRequestTest` and `SignupActivityExceptionTest` to `SignupRepositoryTest` and `SignupRepositoryExceptionTest`, respectively, and moved them to the corresponding `signup` test package.
- Cleaned up test overbuilding by instantiating `SignupRepository` directly with injected dispatcher and mocked dependencies without invoking full Activity via Robolectric reflection, improving test speed and reliability.
- Eliminated dangling, commented-out Toast assertions in exception tests since Toasting behavior remains properly handled by the calling `SignupActivity`.
